Добавил ViewModel на ProductDitaile
